comments addressed. moved yul code generation to keygen
